move Container.tsx to components/
