+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import InMemoryUserRepository from './inMemory/InMemoryUserRepository';
 import Store from './Store';
-import { Container } from './Container/Container';
+import { Container } from './components/Container';
 
 const App = () => {
   InMemoryUserRepository.initializeLocalStorage();
